test(ftn): split lookup tests into nodelist_lookup_test.go
